Jupyter Notebook for Boston Housing
You are the best real estate agent in Boston because you use Machine Learning to predict house prices. One of your clients wants to sell a house and they want you to give them an estimate of the best selling price. You are the Kelley Blue Book of homes, so you want to predict the correct price. Not only that, you should be able to convince your client that your estimate is reasonable.

We have seen a lot of models (DT, NN and so on) and various algorithms (ID3, Gradient Descent, etc.) to learn them in the video lessons. Often when a dataset is given, we are trying to fit it to the best model; that is, the best model that generalizes for data we haven’t seen yet. An important skill of a data scientist is to identify the best model. This project is going to help you analyze a dataset and teach you how to choose the best model that generalizes that dataset.